requirements.txt
# LangGraph / LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# API & HTTP
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-multipart>=0.0.9

# Data
pydantic>=2.7.0
python-dotenv>=1.0.0

# Vector DB
pinecone>=5.0.0

# Monitoring
prometheus-client>=0.20.0

# Web Search
ddgs>=0.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Lambda handler
mangum>=0.17.0